CourseApp
This web application can be created using ASP.NET Core MVC. In the application there will be three main entities: Student, Teacher and Course. Students can enroll in courses, teachers can teach courses, and each course can be seen by which teacher is teaching it and which students are participating.